refactor: Refactor ChatbotInput component to use TypeScript
This commit refactors the ChatbotInput component to use TypeScript instead of JavaScript. It also adds a new prop, multimodeChat, to enable multimode chat functionality. Additionally, it introduces the multimodeRenderFunction prop to allow users to customize the chat mode selector.
